% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMeans.R
\name{colMeans.weighted}
\alias{colMeans.weighted}
\title{Weighted column means}
\usage{
colMeans.weighted(x, w, ...)
}
\arguments{
\item{x}{A numeric matrix or dataframe.}

\item{w}{A vector of weights, must be same length as \code{x}.}

\item{...}{passed to \code{\link{weighted.mean}}}
}
\value{
A vector of weighted means.
}
\description{
Calculate weighted column means of a matrix or dataframe.
}
\examples{
colMeans.weighted(mtcars, runif(nrow(mtcars)))
}

