% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estep.R
\name{estep_cond}
\alias{estep_cond}
\alias{estep_cond2}
\title{Expectation step for Conditional Expectation-Maximisation algorithm}
\usage{
estep_cond(x_A, x_B, pro, mean_A, mean_B, sigma_AA, sigma_AB, sigma_BB, groups)

estep_cond2(x_A, x_B, pro, mean_A, mean_B, sigma_AA, sigma_AB, sigma_BB, groups,
  oldz = NULL)
}
\description{
Similar to the expectation step in a standard E-M algorithm as
  applied to a mixture of Gaussians.
}

