% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{initialise.memberships}
\alias{initialise.memberships}
\title{Initialise parameters for Expectation-Maximisation algorithm}
\usage{
initialise.memberships(x, K)
}
\arguments{
\item{x}{A data frame containing continuous data.}

\item{K}{The number of groups.}
}
\value{
A K x N matrix of 0/1 entries.
}
\description{
Initialise either the cluster membership matrix or the mixture
  model parameters.
}
\examples{
data(mtcars)
initialise.memberships(mtcars, 3)
}

