% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbc.R
\name{mbc}
\alias{mbc}
\title{Fit a mixture of multivariate Gaussians}
\usage{
mbc(x, groups, maxiter = 500)
}
\arguments{
\item{x}{Data frame or a matrix}

\item{groups}{The number of groups/mixture components to fit.}

\item{maxiter}{The maximum number of iterations for the E-M algorithm.}
}
\value{
A list containing the estimated parameters for the mixture
  distribution.
}
\description{
Fit a mixture of multivariate Gaussians
}
\examples{
library(mclust)
data(banknote)
mbc(x = banknote[, -1], groups = 2)
}

