% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.R
\name{mean_conditional}
\alias{mean_conditional}
\alias{sigma_conditional}
\title{Calculate conditional mean and covariance for multivariate Gaussian
  distributions.}
\usage{
mean_conditional(mean, cov, precision_cond, x_cond, mean_cond, sigma_cond)

sigma_conditional(sigma, cov, precision_cond, sigma_cond)
}
\arguments{
\item{mean}{The mean vector for A.}

\item{cov}{The covariance between A and B.}

\item{precision_cond}{The precision matrix for B.}

\item{x_cond}{The value of B on which to condition (only needed for the
mean).}

\item{mean_cond}{The mean vector for B.}

\item{sigma_cond}{The covariance matrix for B, optional alternative to
\code{precision_cond}.}
}
\value{
The relevant parameter for the distribution of A given B.
}
\description{
Calculate conditional mean and covariance for multivariate
  Gaussian distributions. Consider partitioning into A and B. This function
  returns parameters for the distribution of A given B, which is itself
  another Gaussian distribution.
}

