% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{mstep}
\alias{mstep}
\title{Maximisation step for Expectation-Maximisation algorithm}
\usage{
mstep(x, z, groups = NULL, p = NULL)
}
\arguments{
\item{x}{A matrix of data.}

\item{z}{A matrix of cluster memberships/probabilities.}

\item{groups}{Optional, number of groups in the mixture, inferred from
\code{z} if not supplied.}

\item{p}{Optional, dimension of \code{x}, inferred from \code{x} if not
supplied}
}
\value{
A list of parameter estimates for the mixing proportions, mean
  vectors, and covariance matrices.
}
\description{
Maximisation step for Expectation-Maximisation algorithm
}

