% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{mstep_cond}
\alias{mstep_cond}
\title{Conditional maximisation step for Expectation-Maximisation algorithm}
\usage{
mstep_cond(x_A, x_B, mean_A, sigma_AA, sigma_AB = NULL, mean_B = NULL,
  sigma_BB = NULL, pro = NULL, groups, z, likelihood = FALSE,
  method_sigma_AB = c("numeric", "analytic"))
}
\arguments{
\item{x_A}{A matrix of data for previous batch.}

\item{x_B}{A matrix of data for the current batch.}

\item{mean_A}{The mean vectors for the previous batch (considered fixed).}

\item{sigma_AA}{The covariance matrices for the previous batch (considered
fixed).}

\item{sigma_AB}{Starting value for the covariance between batches A and B. If
\code{NULL}, initialised at all zeros.}

\item{mean_B}{Starting value for the mean vectors for the current batch.}

\item{groups}{Optional, number of groups in the mixture, inferred from
\code{z} if not supplied.}

\item{z}{A matrix of cluster memberships/probabilities.}

\item{p}{Optional, dimension of \code{x}, inferred from \code{x} if not
supplied}
}
\value{
A list of parameter estimates for the mixing proportions, mean
  vectors, and covariance matrices.
}
\description{
Conditional maximisation step for Expectation-Maximisation
  algorithm. Maximising paramters for current batch of data conditional on
  the parameters from the previous batch.
}

