% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.wt.R
\name{var.wt}
\alias{var.wt}
\title{Calculate a weighted covariance matrix between two sets of columns}
\usage{
var.wt(x, y = NULL, w, method = c("unbiased", "ML"), xcenter = NULL,
  ycenter = NULL)
}
\arguments{
\item{x}{Data frame or a matrix}

\item{y}{Data frame or a matrix with the same number of rows as \code{x}. If
left \code{NULL}, the covariance matrix of \code{x} is calculated.}

\item{w}{A set of weights.}

\item{method}{The type of estimate returned, \code{"unbiased"} (default) or
\code{"ML"}, meaning maximum likelihood.}
}
\value{
A weighted covariance matrix.
}
\description{
Calculate a weighted covariance matrix in the same way as
  \code{\link{cov.wt}}, but allow two sets of columns as in
  \code{\link{var}}.
}
\examples{
wts <- (tmp <- runif(nrow(mtcars))) / sum(tmp)
var.wt(mtcars[, 1:5], mtcars[, 1:2], w = wts)
cov.wt(mtcars[, 1:5], wt = wts)$cov
}

