% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{calcloglik}
\alias{calcloglik}
\alias{logLik.mbc}
\title{Calculate log-likelihood for a Gaussian mixture model.}
\usage{
calcloglik(x, parameters)

\method{logLik}{mbc}(object, ...)
}
\arguments{
\item{x}{A data frame or matrix of data.}

\item{parameters}{A list of parameters for the model.}

\item{object}{A fitted model object of class \code{"mbc"}.}

\item{...}{not currently used.}
}
\value{
A single numeric value.
}
\description{
Calculate log-likelihood for a Gaussian mixture model.
}

