% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{calcloglik}
\alias{calcloglik}
\alias{calcloglik_split}
\alias{logLik.mbcbigp}
\title{Calculate log-likelihood for a Gaussian mixture model.}
\usage{
calcloglik(x, pro, mean, sigma, groups)

calcloglik_split(x_A, x_B, pro, mean_A, mean_B, sigma_AA, sigma_AB, sigma_BB,
  groups)

\method{logLik}{mbcbigp}(object, ...)
}
\arguments{
\item{x}{A data frame or matrix of data.}

\item{object}{A fitted model object of class \code{"mbc"}.}

\item{...}{not currently used.}

\item{parameters}{A list of parameters for the model.}
}
\value{
A single numeric value.
}
\description{
Calculate log-likelihood for a Gaussian mixture model.
}

