% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{calcloglik}
\alias{calcloglik}
\alias{calcloglik_split}
\alias{logLik.mbcbigp}
\title{Calculate log-likelihood for a Gaussian mixture model.}
\usage{
calcloglik(x, pro, mean, sigma, groups)

calcloglik_split(x_A, x_B, pro, mean_A, mean_B, sigma_AA, sigma_AB, sigma_BB,
  groups)

logLik.mbcbigp(object, ...)
}
\arguments{
\item{x}{A data frame or matrix of data.}

\item{pro}{A vector of mixing proportions.}

\item{mean, mean_A, mean_B}{A 2-D array of mean vectors.}

\item{sigma, sigma_AA, sigma_AB, sigma_BB}{A 3-D array of covariance matrices.}

\item{groups}{The integer number of groups.}

\item{x, x_A, x_B}{A matrix of data.}
}
\value{
A single numeric value.
}
\description{
Calculate log-likelihood for a Gaussian mixture model.
}

