% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estep.R
\name{estep}
\alias{estep}
\title{Expectation step for Expectation-Maximisation algorithm}
\usage{
estep(x, parameters)
}
\arguments{
\item{x}{A matrix of data.}

\item{parameters}{A list containing the parameters of the Gaussian mixture
model.}
}
\value{
A matrix of probabilities of belonging to a cluster.
}
\description{
Expectation step for Expectation-Maximisation algorithm
}

