% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{initialise.memberships}
\alias{initialise.memberships}
\title{Initialise parameters for Expectation-Maximisation algorithm}
\usage{
initialise.memberships(x, groups, method = c("kmeans", "hclust"))
}
\arguments{
\item{x}{A data frame containing continuous data.}

\item{groups}{The number of groups.}

\item{method}{The method to give initial clustering.}
}
\value{
A \code{groups} x \code{nrow(x)} matrix of 0/1 entries denoting cluster
  membership.
}
\description{
Initialise the cluster membership matrix for the
  Expectation-Maximisation algorithm for fitting a mixture of Gaussians.
}
\examples{
data(mtcars)
initialise.memberships(mtcars, 3)
}

