% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbcbigp.R
\name{mbcbigp}
\alias{mbcbigp}
\title{Fit a mixture of multivariate Gaussians using sequential conditioning}
\usage{
mbcbigp(x, groups = 2, batches = 3, batchindex = NULL, maxiter = 10,
  plot = FALSE, likelihood = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{Data frame or a matrix}

\item{groups}{The number of groups/mixture components to fit.}

\item{batches}{The batches to split the columns of \code{x} for
processing. A single number gives the desired number of batches, whereas a
vector of length \code{ncol(x)} is treated as an integer index of batch
membership.}

\item{maxiter}{The maximum number of iterations for the E-M algorithm.}

\item{plot}{Logical, should a visualisation be produced to show progress.}

\item{likelihood}{Logical, should the likelihood be calculated and used as a
stopping criterion.}

\item{verbose}{Logical, should statements about progress be printed.}
}
\value{
A list containing the estimated parameters for the mixture
  distribution.
}
\description{
Fit a mixture of multivariate Gaussians
}
\examples{
\dontrun{
data(banknote, package = "mclust")
mbcbigp(x = banknote[, -1], groups = 2, batches = 2)
mbcbigp(x = banknote[, -1], groups = 2, batches = 3)
}
}

