% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbcbigp.R
\name{mbcbigp}
\alias{mbcbigp}
\title{Fit a mixture of multivariate Gaussians using sequential conditioning}
\usage{
mbcbigp(x, groups, batches = 3, maxiter = 200)
}
\arguments{
\item{x}{Data frame or a matrix}

\item{groups}{The number of groups/mixture components to fit.}

\item{batches}{The number of batches to split the columns of \code{x} for
processing.}

\item{maxiter}{The maximum number of iterations for the E-M algorithm.}
}
\value{
A list containing the estimated parameters for the mixture
  distribution.
}
\description{
Fit a mixture of multivariate Gaussians
}
\examples{
library(mclust)
data(banknote)
mbc(x = banknote[, -1], groups = 2)
}

