% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.R
\name{mean_conditional}
\alias{mean_conditional}
\alias{sigma_conditional}
\alias{dmvnorm_conditional}
\title{Calculate conditional mean and covariance for multivariate Gaussian
  distributions.}
\usage{
mean_conditional(mu_B, mu_A, x_A, sigma_AA, sigma_AB)

sigma_conditional(sigma_BB, sigma_AB, sigma_AA)

dmvnorm_conditional(x_A, x_B, mean_A, mean_B, sigma_AA, sigma_AB, sigma_BB)
}
\arguments{
\item{mu_B}{The mean vector for B.}

\item{mu_A}{The mean vector for A.}

\item{x_A}{The points in A at which to evaluate the conditional mean of B.}

\item{sigma_AA}{The covariance matrix for A.}

\item{sigma_AB}{The cross-covariance matrix for A and B.}
}
\value{
The relevant mean parameter for the distribution of B given A.
}
\description{
Calculate conditional mean and covariance for multivariate
  Gaussian distributions. Consider partitioning into A and B. This function
  returns parameters for the distribution of B given A, which is itself
  another Gaussian distribution.
}

